chore(iterate): relocate the AC15 import-check test to stay under the 300-line file cap
test_rollback_module_imports_no_data_tier_client belongs with
test_rollback.py's other in-process rollback.py unit tests anyway
(no CLI/subprocess harness needed) and moving it there keeps both
files under the bloat gate's 300-line guideline.
