+++ plugins/shipwright-deploy/tests/test_rollback.py
@@ -8,7 +8,9 @@ the returned payload. Clone-strategy and CLI-argument tests live in
 ``test_rollback_clone.py``.
 """
 
+import ast
 import urllib.error
+from pathlib import Path
 
 import pytest
 
@@ -234,4 +236,25 @@ def test_invalid_ref_forms_are_rejected_before_any_host_call(client, bad):
     assert recording.calls == []
     assert result["success"] is False
     assert result["mutated"] is False
-    assert result["halt"] is False
+
+
+@pytest.mark.covers("FR-01.08/AC15")
+def test_rollback_module_imports_no_data_tier_client():
+    """Spec FR-01.08/AC15: "stored data...stays where it is" is an
+    architectural guarantee here, not a runtime one to spy on — makes it
+    machine-checked instead of only docstring-asserted: a future import of
+    a database/migration-execution client would fail this test."""
+    forbidden = {"psycopg2", "psycopg", "sqlalchemy", "asyncpg", "pymysql", "sqlite3", "supabase"}
+    tree = ast.parse(Path(rollback.__file__).read_text(encoding="utf-8"))
+    imported = {
+        alias.name.split(".")[0]
+        for node in ast.walk(tree)
+        if isinstance(node, ast.Import)
+        for alias in node.names
+    } | {
+        node.module.split(".")[0]
+        for node in ast.walk(tree)
+        if isinstance(node, ast.ImportFrom) and node.module
+    }
+    hit = imported & forbidden
+    assert not hit, f"rollback.py must never import a data-tier client, found: {hit}"
